requirements.txt
altgraph==0.17.3
certifi==2022.12.7
charset-normalizer==2.0.12
et-xmlfile==1.1.0
guidata==2.3.1
guiqwt==4.3.1
idna==3.4
Nuitka==1.3.6
numpy==1.24.1
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
peewee==3.15.4
Pillow==9.3.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyqt-tools==1.0.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt6==6.4.0
PyQt6-Qt6==6.4.1
PyQt6-sip==13.4.0
python-blizzardapi==1.3.0
python-dateutil==2.8.2
PythonQwt==0.10.2
pytz==2022.7
QtPy==2.3.0
requests==2.26.0
scipy==1.9.3
six==1.16.0
urllib3==1.26.13
XlsxWriter==3.0.3
